modify main_gui to make class about videoplayer and recorder
